nit #8958

+++ arches/management/commands/updateproject.py
@@ -24,17 +24,33 @@ class Command(BaseCommand):
 
         if os.path.exists(project_webpack_path):
             shutil.rmtree(project_webpack_path)
+
         shutil.copytree(os.path.join(settings.ROOT_DIR, "install", "arches-templates", "project_name", "webpack"), project_webpack_path)
 
-        # publish graphs that were previously active
+        #  check if temp_graph_status table exists
         with connection.cursor() as cursor:
-            cursor.execute("select * from temp_graph_status;")
-            rows = cursor.fetchall()
-            graphs = []
-            for row in rows:
-                graphid = str(row[0])
-                active = row[1]
-                if active:
-                    graphs.append(graphid)
-            management.call_command("graph", operation="publish", update_instances=True, graphs=",".join(graphs))
-            cursor.execute("drop table if exists temp_graph_status")
+            cursor.execute(
+                """
+                SELECT EXISTS (
+                    SELECT FROM 
+                        pg_tables
+                    WHERE 
+                        schemaname = 'public' AND 
+                        tablename  = 'temp_graph_status'
+                );
+                """
+            )
+            result = cursor.fetchone()
+
+            # publish graphs that were previously active
+            if result[0]:  # parses result from row
+                cursor.execute("select * from temp_graph_status;")
+                rows = cursor.fetchall()
+                graphs = []
+                for row in rows:
+                    graphid = str(row[0])
+                    active = row[1]
+                    if active:
+                        graphs.append(graphid)
+                management.call_command("graph", operation="publish", update_instances=True, graphs=",".join(graphs))
+                cursor.execute("drop table if exists temp_graph_status")
